moved method, cleaned up
