+++ wdfn-server/config.py
@@ -100,7 +100,7 @@ if os.getenv('CONTAINER_RUN', False):
     STATIC_ROOT = os.environ.get('STATIC_ROOT', '/static/')
 
 # Mail settings for feedback form
-MAIL_SERVER = 'smtp.usgs.gov'
+MAIL_SERVER = 'smtp1.usgs.gov'
 EMAIL_TARGET = {
     'contact' : 'gs-w-{state_district_code}_NWISWeb_Data_Inquiries@usgs.gov',
     'report': 'gs-w_support_nwisweb@usgs.gov',
