+++ wdfn-server/waterdata/tests/test_utils.py
@@ -10,7 +10,8 @@ import requests as r
 
 from .. import app
 
-from ..utils import construct_url, defined_when, execute_get_request, parse_rdb, set_cookie_for_banner_message
+from ..utils import construct_url, defined_when, execute_get_request, parse_rdb, set_cookie_for_banner_message,\
+    create_message
 
 
 class TestConstructUrl(TestCase):
@@ -32,6 +33,48 @@ class TestConstructUrl(TestCase):
     def test_with_no_params(self):
         expected = 'https://fakeurl.gov/blah1/blah2'
         self.assertEqual(construct_url(self.test_netloc, self.test_path), expected)
+
+
+class TestCreateMessage(TestCase):
+    def test_create_message(self):
+        target_email = 'test@test.com'
+        form_data = {
+            'feedback-type': 'comment',
+            'monitoring-location-url': 'http://localhost:5050/monitoring-location/05413500/',
+            'subject': 'Test Subject',
+            'message': 'test message',
+            'user-email-address': 'test@test.com',
+            'user-name': 'test name',
+            'user-phone-number': 'test number',
+            'user-organization-or-address': 'test organization'
+        }
+        user_system_data = {'Mozilla/5.0': '(X11; Ubuntu; Linux x86_64; rv:84.0) Gecko/20100101 Firefox/84.0)'}
+        timestamp = '2021-02-01 23:55:21.076754'
+        expected = """Subject:
+ User Question/Comment for http://localhost:5050/monitoring-location/05413500/
+From: WDFN Comments and Questions
+Reply-To: test@test.com
+To: test@test.com
+Content-Type: text/plain; charset="utf-8"
+Content-Transfer-Encoding: 7bit
+MIME-Version: 1.0
+
+
+        From: test name
+        Subject: Test Subject
+        Location: http://localhost:5050/monitoring-location/05413500/
+        Time (UTC): 2021-02-01 23:55:21.076754
+        *********** Message ***********
+        test message
+        *********** User Information ***********
+        Email: test@test.com
+        Phone: test number
+        Organization or Address: test organization
+        User Browser/System Details: {'Mozilla/5.0': '(X11; Ubuntu; Linux x86_64; rv:84.0) Gecko/20100101 Firefox/84.0)'}
+        
+"""
+        actual = create_message(target_email, form_data, user_system_data, timestamp)
+        self.assertEqual(expected, str(actual))
 
 
 class TestCookieSetting(TestCase):
